This builds off of @adampalay's work in #7168

Features:
* Max scores cache to limit XModule instantiation to find out the possible score (cache is automatically invalidated when an new version of the course is pushed).
* Filtering to limit the amount of data we pull back in the `FieldDataCache` and tree crawling to only be items that might impact grading.
* Request-level caching of the partitioning service (this was causing many requests for A/B tests)
* Remove many of the StudentModule queries that weren't using the `FieldDataCache`.

On a local test of 8.MechCx (a good stress test, as it has > 800 problems and uses A/B testing more than most), this brings down the progress page render time from 60s (!) to 20s in an uncached state, and 5s in a cached state. Leaning more heavily on `FieldDataCache` means that we go from potentially thousands of CSM queries on particularly large courses, down to one or two. There's clearly more work to do, but this should help curb the grading time on some of our largest courses and significantly reduce CSM usage (about half of CSM usage happens through the progress page).

@cpennington, @adampalay: Review please?