@ormsbee 2015-03-13
@cpennington, @adampalay: I think I've responded to your comments. I didn't do the context manager thing mostly for extra indent. Is that something you want to see before it goes in?

Assuming devops can get the prod replica sorted, I'm going to try some larger scale testing over the weekend. Could you please give :+1: if it's ok to go?