@cpennington 2015-03-13
No, the contextmanager isn't a requirement for merge. I can definitely understand not wanting more nesting