@ormsbee 2015-03-16
This gives about a 3X speed up on load-tests against the demo course. It also gives a lot more regularity: median rendering time went from 750ms to 240ms, but 99th percentile times went from 1400ms to 270ms.

Still running a batch of existing courses through grading runs with old and new to verify that we're producing the same results. If so, this should be able to go out in next week's release cycle.