@ormsbee 2015-03-18
<sigh> There is something about this PR that results in slight course-tree differences for some students in one of the test courses. @nasthagiri: I'm at a conference for the next couple of days. I'm going to try to track this problem down tomorrow night and fix it, but even if I do so, it'll take another day to run the big regression test again. It's unlikely that I'll get to merge this until Friday.