@nasthagiri 2015-03-18
@ormsbee No problem.  In that case, I have separated out the partition cache changes into its own commit and cherry-picked it into my PR: https://github.com/edx/edx-platform/pull/7298/commits
So whoever merges first can commit those changes.