@ormsbee 2015-03-19
Okay, latest update -- I thought it was extra columns because of the length discrepancy, but it was really because we were assigning the score = 0 when it should have been 0.0 (int vs. float). Going to make that change and kick things off on the sandbox again.