@ormsbee 2015-07-08
@cpennington: I fixed the redundant code issue you pointed out. I've rebased, tests pass once again, and if this thing passes and I don't get objections, I plan to hit the merge button tomorrow morning...